Giocatore: Spostato Main. TODO: toString(), Tests.
